+++ src/containers/TransactionPage/TransactionPanel/BookingLocationMaybe.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import classNames from 'classnames';
+
+import { FormattedMessage } from '../../../util/reactIntl';
+
+import AddressLinkMaybe from './AddressLinkMaybe';
+
+import css from './TransactionPanel.module.css';
+
+// Functional component as a helper to build ActivityFeed section
+const BookingLocationMaybe = props => {
+  const { className, rootClassName, listing, showBookingLocation } = props;
+  const classes = classNames(rootClassName || css.bookingLocationContainer, className);
+
+  if (showBookingLocation) {
+    const location = listing?.attributes?.publicData?.location || {};
+    return (
+      <div className={classes}>
+        <h3 className={css.sectionHeading}>
+          <FormattedMessage id="TransactionPanel.bookingLocationHeading" />
+        </h3>
+        <div className={css.bookingLocationContent}>
+          <AddressLinkMaybe
+            linkRootClassName={css.bookingLocationAddress}
+            location={location}
+            geolocation={listing?.attributes?.geolocation}
+            showAddress={true}
+          />
+        </div>
+      </div>
+    );
+  }
+  return null;
+};
+
+export default BookingLocationMaybe;